api/requirements.txt
SQLAlchemy>=1.4.22
Flask>=2.0.1
gunicorn>=20.1.0
requests>=2.25.1
psycopg2-binary>=2.9.1
sqlglot>=25.6.1

llama-index>=0.7.7
weaviate-client>=3.20.0
chromadb>=0.3.29
openai>=0.27.8
python-dotenv>=1.0.0

# Other potential dependencies
#numpy==1.24.3
#pandas==2.0.2